Add routes for GET and POST, render the result.
This application is composed of a front-end and a back-end.

The back-end (express) has two routes:
  - GET /messages
  - POST /messages

The back-end keeps the application state in a file stored at:
  - ./data/message.json

When the back-end receives a POST it updates the ./data/message.json

The front-end requests the current back-end response after an
initial render by React for a first page load.

Before the React form submits the data, react intercepts the request
and manages the exchange with the server. When the response is received
then the React front-end updates its state and passes the state as a
variable to the rest of the page.